so close

+++ disser/Models/EF/GOST/CreatedGOST.cs
@@ -9,7 +9,7 @@ namespace disser.Models.EF.GOST
         public List<RukovoditelWantWork>? RukovoditelWantWork { get; set; }
         public int? ChoisedRukovoditelID { get; set; }
         public bool OnWork { get; set; }
-        public double WorkPercentage { get; set; }
+        public int WorkPercentage { get; set; }
         public string? EndedFile { get; set; }
         public List<TranslateFile>? FullFileToTranslate { get; set; }
         public List<SimilarFile>? similarGOSTs { get; set; }
